+++ packages/framer-motion/src/animation/animators/BaseAnimation.ts
@@ -9,6 +9,7 @@ import {
     AnimationPlaybackControls,
     RepeatType,
     ValueAnimationOptions,
+    ValueAnimationOptionsWithRenderContext,
 } from "../types"
 import { canAnimate } from "./utils/can-animate"
 import { getFinalKeyframe } from "./waapi/utils/get-final-keyframe"
@@ -24,7 +25,7 @@ import { getFinalKeyframe } from "./waapi/utils/get-final-keyframe"
 const MAX_RESOLVE_DELAY = 40
 
 export interface ValueAnimationOptionsWithDefaults<T extends string | number>
-    extends ValueAnimationOptions<T> {
+    extends ValueAnimationOptionsWithRenderContext<T> {
     autoplay: boolean
     delay: number
     repeat: number