+++ packages/framer-motion/src/animation/animators/waapi/utils/style.ts
@@ -1,0 +1,20 @@
+type MotionStyleKey = Exclude<
+    keyof CSSStyleDeclaration,
+    "length" | "parentRule"
+>
+
+export function setCSSVar(
+    element: HTMLElement,
+    name: string,
+    value: string | number
+) {
+    element.style.setProperty(`--${name}`, value as string)
+}
+
+export function setStyle(
+    element: HTMLElement,
+    name: string,
+    value: string | number
+) {
+    element.style[name as MotionStyleKey] = value as any
+}