+++ packages/framer-motion/src/components/LayoutGroup/index.tsx
@@ -7,7 +7,7 @@ import {
     LayoutGroupContextProps,
 } from "../../context/LayoutGroupContext"
 import { DeprecatedLayoutGroupContext } from "../../context/DeprecatedLayoutGroupContext"
-import { nodeGroup } from "../../projection-entry"
+import { nodeGroup } from "../../projection"
 import { useForceUpdate } from "../../utils/use-force-update"
 
 type InheritOption = boolean | "id"