+++ packages/framer-motion/src/value/types/numbers/units.ts
@@ -1,8 +1,6 @@
-import { isString } from "../utils"
-
 const createUnitType = (unit: string) => ({
     test: (v: string | number) =>
-        isString(v) && v.endsWith(unit) && v.split(" ").length === 1,
+        typeof v === "string" && v.endsWith(unit) && v.split(" ").length === 1,
     parse: parseFloat,
     transform: (v: number | string) => `${v}${unit}`,
 })